All computations are now made in double precision.

+++ buffer.h
@@ -27,13 +27,13 @@ public:
 
     int spectrum_num_samples_ = 3;
 
-    std::vector< std::vector< glm::vec3 > > buffer_data_;
+    std::vector< std::vector< glm::dvec3 > > buffer_data_;
 
 private:
 
-    float clamp( float x ) const
+    double clamp( double x ) const
     {
-        return ( x < 0.0f ) ? 0.0f : ( x > 1.0f ) ? 1.0f : x;
+        return ( x < 0.0 ) ? 0.0 : ( x > 1.0 ) ? 1.0 : x;
     }
 };
